Add files via upload
Tradecoin Lotto Dapp is a Waves Blockchain based Dapp. Which will pay users 1.5 Times of their Lotto Size instantly. Tradecoin Dapp is Decentralized and 100% Safe to use. Smart Contract of Dapp is open source and can be deployed any time to cross check and verify the Dapp and payout function.